Dev computation (#5)
* RNG State done, need to test the hashing algorithm to compare results with original function

* DirectionCosine::sample_isotropic done

* DirectionCosine done; need to compare the computed values with original code

* doc

* started to complete macro_cross_section.rs; replaced the usage of a magic value by adding a separate method

* cargo fmt

* completed macro_cross_section functions

* added test to compare results; the pseuo hash function of mc_rng_state gives a different result

* fiex mc_rng_state, the generator now yields the same numbers

* added test for sample_isotropic; added c files used to test original code

* added rotation test; all functions seems to give the same results

* cargo fmt

* computation test can be run using --nocapture to compare results with the original functions

* update_trajectory() done

* finished collision_event

* cargo fmt

* finished segment_outcome

* cargo fmt

* added rust test for update trajectory & needed c++ files

* wrote test in c files; need to add a makefile and check results

* velocity & direction cosine differ

* corrected test, update_trajectory ok

* very simple makefile

* added internal functions of mct.rs; this module needs refactor asap

* started implementation of mct.rs function; lots of refactoring to do on this file

* completed generate_coordinate_3dg; updating functions signature

* cargo fmt

* finished public mct functions; cleanup needed

* cargo fmt

* mct functions continued; will cleanup after mct & facet_crossing_event are completed

* mct completed; cleanup todo

* completed facet_crossing_event

* cargo fmt

* mct cleanup 1/?

* mct cleanup 2/2; updated MCLocation to avoid magic values

* cargo fmt

* added test for volume computation & move particle

* all test added; the belongs_or_return macro has diverging results

* test passed; fixed a mistake in the C code

* cleanup & format

+++ src/mc/mc_facet_adjacency.rs
@@ -32,7 +32,7 @@ impl Default for SubfacetAdjacency {
 #[derive(Debug)]
 pub struct MCFacetAdjacency {
     pub subfacet: SubfacetAdjacency,
-    pub num_points: u32,
+    pub num_points: usize,
     pub point: [u32; 3],
 }
 
@@ -46,7 +46,7 @@ impl Default for MCFacetAdjacency {
 #[derive(Debug)]
 pub struct MCFacetAdjacencyCell {
     pub facet: Vec<MCFacetAdjacency>,
-    pub point: Vec<u32>,
+    pub point: Vec<usize>,
 }
 
 impl Default for MCFacetAdjacencyCell {
